# Use `outlines.processors` and `SequenceGeneratorAdapter` for `outlines.models.vllm`