--- a/outlines/generate/choice.py
+++ b/outlines/generate/choice.py
@@ -1,7 +1,7 @@
 from functools import singledispatch
 from typing import Callable, List
 
-from outlines.generate.api import SequenceGenerator
+from outlines.generate.api import SequenceGeneratorAdapter
 from outlines.models import OpenAI
 from outlines.samplers import Sampler, multinomial
 
@@ -11,7 +11,7 @@ from .regex import regex
 @singledispatch
 def choice(
     model, choices: List[str], sampler: Sampler = multinomial()
-) -> SequenceGenerator:
+) -> SequenceGeneratorAdapter:
     regex_str = r"(" + r"|".join(choices) + r")"
 
     generator = regex(model, regex_str, sampler)