--- a/outlines/generate/format.py
+++ b/outlines/generate/format.py
@@ -1,7 +1,7 @@
 from functools import singledispatch
 
 from outlines.fsm.types import python_types_to_regex
-from outlines.generate.api import SequenceGenerator
+from outlines.generate.api import SequenceGeneratorAdapter
 from outlines.models import OpenAI
 from outlines.samplers import Sampler, multinomial
 
@@ -9,7 +9,9 @@ from .regex import regex
 
 
 @singledispatch
-def format(model, python_type, sampler: Sampler = multinomial()) -> SequenceGenerator:
+def format(
+    model, python_type, sampler: Sampler = multinomial()
+) -> SequenceGeneratorAdapter:
     """Generate structured data that can be parsed as a Python type.
 
     Parameters