--- a/outlines/processors/base_logits_processor.py
+++ b/outlines/processors/base_logits_processor.py
@@ -67,22 +67,18 @@ class OutlinesLogitsProcessor(Protocol):
         3) Call self.process_logits() to perform business logic
         4) Cast logits back to original array library type
         """
-
         # ensure logits are torch Tensors
         torch_logits = self._to_torch(logits)
+        input_ids = self._to_torch(input_ids)
 
-        assert torch_logits.shape[:-1] == self._to_torch(input_ids).shape[:-1]
-
-        # ensure input_ids are List
-        if not isinstance(input_ids, list):
-            input_ids = input_ids.tolist()  # compatible with numpy, torch, and mlx
+        assert torch_logits.shape[:-1] == input_ids.shape[:-1]
 
         # Guarantee passed as 2D Tensors, then covert back to original (1D or 2D) shape
         if len(torch_logits.shape) == 2:
-            processed_logits = self.process_logits(input_ids, torch_logits)
+            processed_logits = self.process_logits(input_ids.tolist(), torch_logits)
         elif len(torch_logits.shape) == 1:
             processed_logits = self.process_logits(
-                [input_ids], torch_logits.unsqueeze(0)
+                [input_ids.tolist()], torch_logits.unsqueeze(0)
             ).squeeze(0)
 
         # return logits as passed array type
@@ -97,7 +93,7 @@ class OutlinesLogitsProcessor(Protocol):
         elif isinstance(tensor_like, np.ndarray):
             return torch.from_numpy(tensor_like)
 
-        elif isinstance(tensor_like, list):
+        elif isinstance(tensor_like, (list, tuple)):
             return torch.tensor(tensor_like)
 
         elif is_mlx_array_type(type(tensor_like)):
@@ -108,7 +104,8 @@ class OutlinesLogitsProcessor(Protocol):
         else:
             raise TypeError(
                 "LogitsProcessor must be called with either np.NDArray, "
-                "torch.Tensor, list, or mlx.core.array typed logits"
+                "torch.Tensor, list, or mlx.core.array typed logits. "
+                f"Logits type: `{type(tensor_like)}`"
             )
 
     @staticmethod
@@ -122,6 +119,9 @@ class OutlinesLogitsProcessor(Protocol):
 
         elif target_type == list:
             return tensor.detach().tolist()
+
+        elif target_type == tuple:
+            return tuple(tensor.detach().tolist())
 
         elif is_mlx_array_type(target_type):
             import mlx.core as mx